Extract MainHeader into separate file
