+++ lib/annotate_rb/model_annotator/annotation.rb
@@ -4,6 +4,7 @@ module AnnotateRb
   module ModelAnnotator
     module Annotation
       autoload :AnnotationBuilder, "annotate_rb/model_annotator/annotation/annotation_builder"
+      autoload :MainHeader, "annotate_rb/model_annotator/annotation/main_header"
     end
   end
 end
