+++ lib/annotate_rb/model_annotator/annotation/annotation_builder.rb
@@ -4,37 +4,6 @@ module AnnotateRb
   module ModelAnnotator
     module Annotation
       class AnnotationBuilder
-        class MainHeader < Components::Base
-          # Annotate Models plugin use this header
-          PREFIX = "== Schema Information"
-          PREFIX_MD = "## Schema Information"
-
-          attr_reader :version
-
-          def initialize(version, include_version)
-            @version = version
-            @include_version = include_version
-          end
-
-          def to_markdown
-            header = "# #{PREFIX_MD}"
-            if @include_version && version > 0
-              header += "\n# Schema version: #{version}"
-            end
-
-            header
-          end
-
-          def to_default
-            header = "# #{PREFIX}"
-            if @include_version && version > 0
-              header += "\n# Schema version: #{version}"
-            end
-
-            header
-          end
-        end
-
         END_MARK = "== Schema Information End"
 
         MD_NAMES_OVERHEAD = 6
